add an eBPF program to track CPU usage in softirq processing
A new eBPF program has been created to monitor the CPU usage for packet
processing within the Linux softirq NET_TX and NET_RX. This program
consists of two parts: eBPF probes attached to softirq_entry and
softirq_exit, and an userspace program responsible for attaching the
probes and retrieving the CPU usage data evaluated using the probes in
softirq context.

The eBPF probe attached to softirq_entry records the time when packet
processing begins in NET_TX and/or NET_RX, storing this information in
an eBPF map. The eBPF probe attached to softirq_exit registers the
timestamp when packet processing is completed, allowing for the
calculation of the amount of time the system spends processing each
packet.

This new eBPF program provides valuable insights into CPU usage during
packet processing, allowing for more efficient monitoring and
optimization of network performance.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>